combase: fix race condition in CoWaitForMultipleHandles timeout for MTA.
There is a potential race condition for very small timeouts,
as GetTickCount() could increment between `start_time` and the
first update of `now`, causing the loop to immediately return timeout
(RPC_S_CALLPENDING) without having even checked the handles.

The !message_loop else has no reason to do this; it only runs once,
and then always exits via the `break` statement. The only `continue` that
can cause the loop to iterate again is in the messages available branch
In an the one-and-only call to WaitForMultipleObjectsEx should simply get
the entire specified timeout.

The message_loop (STA) case has the same race condition, but it will be
fixed in a later commit (after some preparatory refactoring).