combase: consolidate handling of QS_ALLPOSTMESSAGE into com_peek_message.
The order of operations here is quite sensitive, in order to keep the
desired priority of various message types and queue state in sync.
Splitting it betwen two functions makes the dependencies less obvious.
Combining them into a single if/else makes it clearer that both are ways
to peek that will clear QS_ALLPOSTMESSAGE so MsgWaitForMultipleObjects
waits. When there is an apt->win, the message is removed and pumped;
without one only the side-effect (clearing QS_ALLPOSTMESSAGE) is wanted.