combase: CoWaitForMultipleHandles needs to check completion before messages.
All cases, not just COWAIT_ALTERTABLE (APCs) should check for success
before blocking in MsgWaitForMultipleObjectsEx. Using COWAIT_WAITALL
in an STA has the same well-known drawback as MWMO_WAITALL; the queue
status then counts as part of the "all handles" return.

But this test shows that that should apply only if there is a need to
wait for messages at all - if all the specified handles are signaled,
CoWaitForMultipleHandles shouldn't check messages at all.