shell32/tests: Add tests for drag and drop helpers
Add tests for many possible situations in which the functions defined by
helper interfaces may be used. No tests for rendering the drag image
itself are currently being done. The objective is to allow for a minimal
implementation which could allow dragging to succeed, as there are
situations in which the application tests for success from the drag and drop
helpers before attempting the dragging operation.

The tests contain a mock IDataObject implementation (only GetData() and
SetData() are used, and they are not fully implemented) and a window
procedure to test the message DI_GETDRAGIMAGE along with the tests done
on the function arguments and function call ordering, among others.

Signed-off-by: Tarcísio Ladeia de Oliveira <wyrquill@gmail.com>