Implements basic rpcrt4.dll.RpcServerUseAllProtseqsIf
A very simple function. Normally rpc servers are handed endpoints one at a time, but the app has the option of associating one or more endpoints and protocol pairs to a given interface object, and then asking the rpc server to use them all. This replaces a stub that had caused the entire wineserver to crash (abort).