gitlab: Install libpcsclite-dev and pcscd for winscard.
Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=54656
---
libpcsclite-dev:i386 should not be needed for 32-bit support, but
installing it should not be an issue either.
This still does not provide a virtual smartcard so most of the tests
will still be skipped. But at least it allows to reproduce bug 54661.