advapi32/tests: Replace the custom loop with the tryok() mechanism.
The custom loop was meant to work around the race condition with
services being started / stopped by Windows while the test is running.
The tryok() mechanism provides the same functionality in a simpler to
use and standardized form.
Also this is really only needed on Windows so use
LOOP_ON_FLAKY_WINDOWS_TESTS().