winewayland.drv: Add unaccelerated pointer support
Adds the registry key
HKEY_CURRENT_USER\\Software\\Wine\\Wayland Driver\\unaccelerated_pointer
witch allows raw input. This makes it easier
to calculate the same sensitivity in different games,
use sensitivity calculators, and easily change values
when changing DPI and do not depend on the compositor or OS.

For example, you want to set the sensitivity to half as much,
and sensitivity curves in libinput are more difficult
to calculate than in the games themselves.