winegstreamer: Use video info stride in buffer meta rather than videoflip.
This removes the need for an extra videoconvert, and any ambiguity that
comes with it. Also reduce our dependency on the GStreamer elements to
behave as we assume they do, and reduce the chances to see unnecessary
copies happening which can be harmful with large buffers. Some GStreamer
versions for instance have a bogus videoflip that causes a buffer copy
to happen when it is supposed to be pass through.