winegstreamer: Avoid an assert in Gstreamer code.
I have seen pipelines with a vaapipostproc element, which makes the caps non-fixed. wg_format_from_caps calls
gst_audio_info_from_caps, which expects fixed caps and asserts if they are not fixed. This does not result in an error
in release builds, but we should still avoid it.