jscript: Use DISP_E_UNKNOWNNAME when looking up a prop that doesn't exist.
This is necessary for Typed Arrays, because they need to avoid allocating
a prop when an indexed prop is used (not just for optimization purposes,
but it's actually required for correctness per the spec), but must also
allow the caller to know when the resulting index is "out of bounds". In
other words, a Typed Array overrides every positive index lookup, even
if it's larger than the Typed Array's size! But it does act like the prop
doesn't exist in those circumstances, so its actual bounds are still needed.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>